requirements.txt
openai
python-telegram-bot
requests
pymysql
pandas
reportlab
yfinance
ibapi
